pyproject.toml
[project]
name = "bookstore"
version = "0.1.0"
description = "Bookstore API"
authors = [
    {name = "tiago-silva-batista",email = "tiagosbsantos@outlook.comm"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytest (>=8.4.1,<9.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "django (>=5.2.5,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-rest-framework (>=0.1.0,<0.2.0)",
    "django-extensions (>=4.1,<5.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
